Refactor: OOTD 추천 시 API 응답 스펙 맞지 않는 부분 수정, S3 presined URL 응답하도록 DTO 수정
