+++ src/main/java/project/closet/feed/service/basic/BasicFeedService.java
@@ -273,8 +273,8 @@ public class BasicFeedService implements FeedService {
     private FeedDto toFeedDto(Feed feed, long commentCount, boolean likedByMe) {
         List<OotdDto> ootdDtos = feed.getFeedClothesList().stream()
                 .map(feedClothes -> {
-                    s3ContentStorage.getPresignedUrl(feedClothes.getClothes().getImageKey());
-                    return OotdDto.from(feedClothes.getClothes());
+                    String presignedUrl = s3ContentStorage.getPresignedUrl(feedClothes.getClothes().getImageKey());
+                    return OotdDto.from(feedClothes.getClothes(), presignedUrl);
                 })
                 .toList();
         String presignedUrl =
